checkpoint

+++ scopehal/DSLabsOscilloscope.cpp
@@ -467,7 +467,7 @@ bool DSLabsOscilloscope::AcquireData()
 
 	//Process analog captures in parallel
 	#pragma omp parallel for
-	for(size_t i=0; i<awfms.size(); i++)
+	for(int64_t i=0; i<awfms.size(); i++)
 	{
 		auto cap = awfms[i];
 		cap->PrepareForCpuAccess();
