+++ scopeprotocols/ThresholdFilter.cpp
@@ -106,7 +106,7 @@ void ThresholdFilter::Refresh()
 		if(hys == 0)
 		{
 			#pragma omp parallel for
-			for(size_t i=0; i<len; i++)
+			for(int64_t i=0; i<len; i++)
 				cap->m_samples[i] = sdin->m_samples[i] > midpoint;
 		}
 		else
@@ -139,7 +139,7 @@ void ThresholdFilter::Refresh()
 		if(hys == 0)
 		{
 			#pragma omp parallel for
-			for(size_t i=0; i<len; i++)
+			for(int64_t i=0; i<len; i++)
 				cap->m_samples[i] = udin->m_samples[i] > midpoint;
 		}
 		else
